backend/requirements.txt
fastapi
uvicorn
pandas
supabase
pydantic
python-dotenv
